Move Icon under ./components
