+++ src/ui/components/icons/icon-cart.svelte
