+++ src/ui/components/icons/icon-error.svelte
