+++ src/ui/components/icons/icon-lock.svelte
