+++ src/ui/components/icons/icon-success.svelte
